Fix state being saved between shader variations
Fixed a bug where instantiated ParamGradientControl nodes would save the state (namely position and the ColourIndicator from the previous shader. I.e opening the gradient and then switching to a new preset.

These nodes are freed and rebuilt now when changing variations of the same shader to ensure they have an up to date state