fix #4 by fully cleaning up RIDs
Objects created in the rendering device need to be manually cleaned up, I had not freed the RIDs for the shader itself and the uniforms.

This seems to be fixed now, however I'll need to keep an eye on it in case there are anymore floating RIDS or dependancies i've not found.