Add support for Single Material category
Adds support for a new category of material called 'single'.

The realistic and stylised PBR shaders in ProcTile are often combinations of materials, patterns and layers to create a larger / more complete texture. Examples might include a brick wall which has both clay bricks, and a sand and cement mortar, as well as the pattern.

Single textures aim to just be the underlying material that the user can then use more widely, such as clay, sand, woodgrain, carbon etc, without being shoehorned into a specific usecase.