Fix broken tiling from hash changes
Some of the adjustments to hashing led to tiling breaking, primarily on the generation of fbm_perlin.

I'm unsure if this is an issue with float seeds being converted to a uint for murmur and having low entropy, or some other issue.

This links to a wider examination of hashing discussed in #18