Add additional safety around file loading and RCE
It is possible to RCE through .cfg files. I have further mitigation planned as well as the features that will hopefully enter the engine in the future, but this is a speedbump at the very least to the script kiddies who might try something.

Also fixed a small bug where the material at index 0 would be considered false and return a loading error.